Rename some spec files I forgot to
