+++ spec/identity/hostdata/ec2_spec.rb
