+++ spec/identity/hostdata/s3_spec.rb
