+++ spec/identity/login_gov/hostdata_spec.rb
@@ -1,0 +1,8 @@
+RSpec.describe 'backwards-compatible name' do
+  it 'exists' do
+    expect do
+      require 'login_gov/hostdata'
+      LoginGov::Hostdata.env
+    end.to_not raise_error
+  end
+end
